after integrating migrate

+++ backend/flaskr/__init__.py
@@ -1,7 +1,5 @@
 
 import os
-from re import A
-from unicodedata import category
 from flask import Flask, request, abort, jsonify, session
 from flask_sqlalchemy import SQLAlchemy
 from flask_cors import CORS
@@ -35,15 +33,19 @@ def create_app(test_config=None):
     Create an endpoint to handle GET requests
     for all available categories.
     """
-    @app.route('/api/categories')
+    @app.route('/api/categories', methods=['GET'])
     def all_categories():
-        categories = Category.query.all()
-        formatted_categories = {category.format()['id']:category.format()['type'] for category in categories}
-        session['current_category'] = 0
-        return jsonify({
-            'success': True,
-            'categories': formatted_categories
-        }), 200
+        try:
+            categories = Category.query.all()
+            formatted_categories = {category.format()['id']:category.format()['type'] for category in categories}
+            session['current_category'] = 0
+
+            return jsonify({
+                'success': True,
+                'categories': formatted_categories
+            }), 200
+        except:
+            abort(422)
 
 
     """
@@ -64,20 +66,29 @@ def create_app(test_config=None):
         start_index = (page - 1) * QUESTIONS_PER_PAGE
         end_index = start_index + QUESTIONS_PER_PAGE
 
-        questions = Question.query.all()
-        formatted_questions = [question.format() for question in questions]
+        try:
+            questions = Question.query.all()
+            formatted_questions = [question.format() for question in questions]
 
-        categories = Category.query.all()
-        formatted_categories = {category.format()['id']:category.format()['type'] for category in categories}
+            categories = Category.query.all()
+            formatted_categories = {category.format()['id']:category.format()['type'] for category in categories}
+        except:
+            abort(422)
+        #reset current category
         session['current_category'] = 0
+
+        current_questions = formatted_questions[start_index: end_index]
+        if len(current_questions) == 0:
+            abort(404, 'page not found')
         
         return jsonify({
             'success': True,
-            'questions': formatted_questions[start_index: end_index],
+            'questions': current_questions,
             'total_questions': len(formatted_questions),
-            'current_category': 'active category',
+            'current_category': session['current_category'],
             'categories': formatted_categories
-        })
+        }), 200
+
 
 
     """
@@ -91,10 +102,16 @@ def create_app(test_config=None):
     @app.route('/api/questions/<int:question_id>', methods=['DELETE'])
     def delete_question(question_id):
         question = Question.query.get(question_id)
-        question.delete()
-        return jsonify({
-            'success': True
-        }), 200
+        print(question)
+        if question is None:
+            abort(404, description='question not found')
+        try:
+            question.delete()
+            return jsonify({
+                'success': True
+            }), 200
+        except:
+            abort(422)
 
     """
     @TODO:
@@ -129,7 +146,7 @@ def create_app(test_config=None):
                     questions = Question.query.filter(Question.question.ilike(f'%{search_term}%')).all()
 
                 formatted_questions = [question.format() for question in questions]
-
+                print(len(formatted_questions))
                 return jsonify({
                     'success': True,
                     'questions': formatted_questions,
@@ -141,7 +158,8 @@ def create_app(test_config=None):
                 new_question = Question(**request.get_json())
                 new_question.insert()
                 return jsonify({
-                    'success': True
+                    'success': True,
+                    'question': new_question.question
                 }), 200
 
         except:
@@ -155,17 +173,26 @@ def create_app(test_config=None):
     categories in the left column will cause only questions of that
     category to be shown.
     """
-    @app.route('/api/categories/<int:category_id>/questions')
+    @app.route('/api/categories/<int:category_id>/questions', methods=['GET'])
     def category_questions(category_id):
-        questions = Question.query.filter_by(category=category_id).all()
-        formatted_questions = [question.format() for question in questions]
-        session['current_category'] = category_id
-        return jsonify({
-            'success': True,
-            'questions': formatted_questions,
-            'total_questions': len(formatted_questions),
-            'current_category': Category.query.get(category_id).type
-        })
+        category = Category.query.get(category_id)
+
+        if category is None:
+            abort(404, description='category not found')
+        
+        try:
+            questions = Question.query.filter_by(category=category_id).all()
+            formatted_questions = [question.format() for question in questions]
+            session['current_category'] = category_id
+
+            return jsonify({
+                'success': True,
+                'questions': formatted_questions,
+                'total_questions': len(formatted_questions),
+                'current_category': category.type
+            })
+        except:
+            abort(422)
 
     """
     @TODO:
@@ -182,33 +209,51 @@ def create_app(test_config=None):
     @app.route('/api/quizzes', methods=['POST'])
     def quiz():
         question = {}
-        previous_questions_id = request.get_json().get('previous_questions', [])
-        quiz_category = request.get_json().get('quiz_category', 0).get('id', 0)
+        quiz_category = 0
+        previous_questions_id = ''
+
+        try:
+            quiz_category = request.get_json().get('quiz_category', 0).get('id', 0)
+            previous_questions_id = request.get_json().get('previous_questions', [])
+        except:
+            abort(400)
+
+        if type(previous_questions_id) != list:
+            abort(422)
+
+        category_questions = []
 
         if quiz_category == 0:
             category_questions = Question.query.all() #handles 'all' category request
         else:
-            category_questions = Question.query.filter_by(category=quiz_category).all()
+            category_questions = Question.query.filter_by(category=quiz_category).all() #handles specific category request
+
+        if not category_questions:
+            abort(404, description='no questions or category does not exist') 
 
         category_questions_id = [question.id for question in category_questions]
-        next_question_id = random.choice(category_questions_id)
 
+        #make random question id
+        next_question_id = random.choice(category_questions_id)
         if len(previous_questions_id) < len(category_questions):
             while next_question_id in previous_questions_id:
                 next_question_id = random.choice(category_questions_id)
         else:
             next_question_id = None
 
+        #get question with matching id
         if next_question_id is not None:
             for q in category_questions: #avoids another interaction with db
                 if q.id == next_question_id:
                     question = q.format()
                     break
+
             return jsonify({
                 'success': True,
                 'question': question
             }), 200
         else:
+            #for end of quiz
             return jsonify({
                 'success': True,
             }), 200
@@ -219,6 +264,62 @@ def create_app(test_config=None):
     Create error handlers for all expected errors
     including 404 and 422.
     """
+
+    @app.errorhandler(404)
+    def not_found(error):
+        return jsonify(
+                {
+                    "success": False,
+                    "error": 404, 
+                    "message": error.description
+                }
+            ), 404
+        
+
+    @app.errorhandler(405)
+    def not_allowed(error):
+        return jsonify(
+                {
+                    "success": False, 
+                    "error": 405, 
+                    "message": "method not allowed"
+                }
+            ), 405
+
+
+    @app.errorhandler(422)
+    def unprocessable(error):
+        return jsonify(
+                {
+                    "success": False, 
+                    "error": 422, 
+                    "message": "unprocessable entity"
+                }
+            ), 422
+
+
+    @app.errorhandler(400)
+    def bad_request(error):
+        return jsonify(
+                {
+                    "success": False, 
+                    "error": 400, 
+                    "message": "bad request"
+                }
+            ), 400
+
+
+    @app.errorhandler(500)
+    def not_allowed(error):
+        return jsonify(
+                {
+                    "success": False, 
+                    "error": 500, 
+                    "message": "something went wrong"
+                }
+            ), 500
+
+
 
     return app
